+++ source/common/upstream/load_balancer_impl.h
@@ -691,56 +691,16 @@ class LoadBalancerShuffleSubsetInfoImpl : public LoadBalancerShuffleSubsetInfo {
 public:
   LoadBalancerShuffleSubsetInfoImpl(
       const envoy::config::cluster::v3::Cluster::LbShuffleSubsetConfig& subset_config)
-      : // enabled_(!subset_config.subset_selectors().empty()),
+      : shard_size_(PROTOBUF_GET_WRAPPED_OR_DEFAULT(subset_config, shard_size, 0)) { }
 
-        shard_size_(
-          PROTOBUF_GET_WRAPPED_OR_DEFAULT(subset_config, shard_size, 0))
-
-        // shard_size_(subset_config.shard_size()),// fallback_policy_(subset_config.fallback_policy()),
-      //  default_subset_(subset_config.default_subset()),
-      //  locality_weight_aware_(subset_config.locality_weight_aware()),
-      //  scale_locality_weight_(subset_config.scale_locality_weight()),
-      //  panic_mode_any_(subset_config.panic_mode_any()), list_as_any_(subset_config.list_as_any()) {
-      {
-    // for (const auto& subset : subset_config.subset_selectors()) {
-    //   if (!subset.keys().empty()) {
-    //     subset_selectors_.emplace_back(std::make_shared<ShuffleSubsetSelectorImpl>(
-    //         subset.keys(), subset.fallback_policy(), subset.fallback_keys_subset(),
-    //         subset.single_host_per_subset()));
-    //   }
-    // }
-  }
-
-  // Upstream::LoadBalancerShuffleSubsetInfo
   bool isEnabled() const override { return shard_size_ > 0; }
-  // envoy::config::cluster::v3::Cluster::LbShuffleSubsetConfig::LbSubsetFallbackPolicy
-  // fallbackPolicy() const override {
-  //   return fallback_policy_;
-  // }
+
   uint32_t shard_size() const override {
     return shard_size_;
   }
 
-  // const ProtobufWkt::Struct& defaultSubset() const override { return default_subset_; }
-  // const std::vector<ShuffleSubsetSelectorPtr>& subsetSelectors() const override {
-  //   return subset_selectors_;
-  // }
-  // bool localityWeightAware() const override { return locality_weight_aware_; }
-  // bool scaleLocalityWeight() const override { return scale_locality_weight_; }
-  // bool panicModeAny() const override { return panic_mode_any_; }
-  // bool listAsAny() const override { return list_as_any_; }
-
 private:
-  // const bool enabled_;
   const uint64_t shard_size_;
-  // const envoy::config::cluster::v3::Cluster::LbShuffleSubsetConfig::LbSubsetFallbackPolicy
-  //     fallback_policy_;
-  // const ProtobufWkt::Struct default_subset_;
-  // std::vector<ShuffleSubsetSelectorPtr> subset_selectors_;
-  // const bool locality_weight_aware_;
-  // const bool scale_locality_weight_;
-  // const bool panic_mode_any_;
-  // const bool list_as_any_;
 };
 
 } // namespace Upstream
