+++ source/common/upstream/subset_lb.h
@@ -275,282 +275,5 @@ private:
   friend class SubsetLoadBalancerDescribeMetadataTester;
 };
 
-class ShuffleSubsetLoadBalancer : public LoadBalancer, Logger::Loggable<Logger::Id::upstream> {
-public:
-  ShuffleSubsetLoadBalancer(
-      LoadBalancerType lb_type, PrioritySet& priority_set, const PrioritySet* local_priority_set,
-      ClusterStats& stats, Stats::Scope& scope, Runtime::Loader& runtime,
-      Random::RandomGenerator& random, const LoadBalancerShuffleSubsetInfo& subsets,
-      const absl::optional<envoy::config::cluster::v3::Cluster::RingHashLbConfig>&
-          lb_ring_hash_config,
-      const absl::optional<envoy::config::cluster::v3::Cluster::MaglevLbConfig>& lb_maglev_config,
-      const absl::optional<envoy::config::cluster::v3::Cluster::LeastRequestLbConfig>&
-          least_request_config,
-      const envoy::config::cluster::v3::Cluster::CommonLbConfig& common_config);
-  ~ShuffleSubsetLoadBalancer() override;
-
-  // Upstream::LoadBalancer
-  HostConstSharedPtr chooseHost(LoadBalancerContext* context) override;
-  // TODO(alyssawilk) implement for non-metadata match.
-  HostConstSharedPtr peekAnotherHost(LoadBalancerContext*) override { return nullptr; }
-
-private:
-  using HostPredicate = std::function<bool(const Host&)>;
-  struct SubsetSelectorFallbackParams;
-
-  std::vector<uint32_t>* combo(uint32_t i, uint32_t n, uint32_t k);
-  // void initSubsetAnyOnce();
-  // void initSubsetSelectorMap();
-  // void initSelectorFallbackSubset(const envoy::config::cluster::v3::Cluster::LbShuffleSubsetConfig::
-  //                                     LbSubsetSelector::LbShuffleSubsetSelectorFallbackPolicy&);
-  // HostConstSharedPtr
-  // chooseHostForSelectorFallbackPolicy(const SubsetSelectorFallbackParams& fallback_params,
-  //                                     LoadBalancerContext* context);
-
-  // Represents a subset of an original HostSet.
-  class HostSubsetImpl : public HostSetImpl {
-  public:
-    HostSubsetImpl(const HostSet& original_host_set) //, bool locality_weight_aware)//, _locality_weight_aware_
-                   // bool scale_locality_weight)
-        : HostSetImpl(original_host_set.priority(), original_host_set.overprovisioningFactor()),
-        original_host_set_(original_host_set) {}
-          // original_host_set_(original_host_set), locality_weight_aware_(locality_weight_aware) {} // _locality_weight_aware_
-          // scale_locality_weight_(scale_locality_weight) {}
-
-    // HostSubsetImpl(const HostSubsetImpl &old_obj) :
-    //   HostSetImpl(old_obj.original_host_set_.priority(), old_obj.original_host_set_.overprovisioningFactor()), original_host_set_(old_obj.original_host_set_),
-    //   locality_weight_aware_(old_obj.locality_weight_aware_) {} //, scale_locality_weight_(old_obj.scale_locality_weight_) {}
-    //
-
-    // HostSubsetImpl(uint32_t priority, uint32_t overprovisioning_factor,
-    //       HostVectorConstSharedPtr hosts, HealthyHostVectorConstSharedPtr healthy_hosts,
-    //       DegradedHostVectorConstSharedPtr degraded_hosts, ExcludedHostVectorConstSharedPtr excluded_hosts)
-    //     : HostSetImpl(priority, overprovisioning_factor,
-    //           hosts, healthy_hosts,
-    //           degraded_hosts, excluded_hosts) {}
-
-
-    // HostSetImplPtr shuffle();
-    // void shuffle();
-
-    void update(const HostVector& hosts_added, const HostVector& hosts_removed,
-                HostPredicate predicate);
-    // LocalityWeightsConstSharedPtr
-    // determineLocalityWeights(const HostsPerLocality& hosts_per_locality) const;
-
-  private:
-    const HostSet& original_host_set_;
-    // const bool locality_weight_aware_;
-    // const bool scale_locality_weight_;
-  };
-
-  // Represents a subset of an original PrioritySet.
-  class PriorityShuffleSubsetImpl : public PrioritySetImpl {
-  public:
-    PriorityShuffleSubsetImpl(const ShuffleSubsetLoadBalancer& subset_lb, HostPredicate predicate);//,
-                      // bool locality_weight_aware); // _scale_locality_weight_
-
-    void update(uint32_t priority, const HostVector& hosts_added, const HostVector& hosts_removed);
-
-    // bool empty() { return empty_; }
-
-    // PriorityShuffleSubsetImpl shuffle(int i);
-
-    void triggerCallbacks() {
-      for (size_t i = 0; i < hostSetsPerPriority().size(); ++i) {
-        runReferenceUpdateCallbacks(i, {}, {});
-      }
-    }
-
-    void updateSubset(uint32_t priority, const HostVector& hosts_added,
-                      const HostVector& hosts_removed, HostPredicate predicate) {
-      reinterpret_cast<HostSubsetImpl*>(host_sets_[priority].get())
-          ->update(hosts_added, hosts_removed, predicate);
-
-      runUpdateCallbacks(hosts_added, hosts_removed);
-    }
-
-    // Thread aware LB if applicable.
-    ThreadAwareLoadBalancerPtr thread_aware_lb_;
-    // Current active LB.
-    LoadBalancerPtr lb_;
-
-  protected:
-    HostSetImplPtr createHostSet(uint32_t priority,
-                                 absl::optional<uint32_t> overprovisioning_factor) override;
-
-  private:
-    const PrioritySet& original_priority_set_;
-    const HostPredicate predicate_;
-    // const bool locality_weight_aware_;
-    // const bool scale_locality_weight_;
-    // bool empty_ = true;
-  };
-
-  using HostSubsetImplPtr = std::shared_ptr<HostSubsetImpl>;
-  using PriorityShuffleSubsetImplPtr = std::shared_ptr<PriorityShuffleSubsetImpl>;
-
-  // using SubsetMetadata = std::vector<std::pair<std::string, ProtobufWkt::Value>>;
-
-  class LbShuffleSubsetEntry;
-  struct SubsetSelectorMap;
-
-  // using LbShuffleSubsetEntryPtr = std::shared_ptr<LbShuffleSubsetEntry>;
-  using SubsetSelectorMapPtr = std::shared_ptr<SubsetSelectorMap>;
-  // using ValueSubsetMap = absl::node_hash_map<HashedValue, LbShuffleSubsetEntryPtr>;
-  // using LbSubsetMap = absl::node_hash_map<std::string, ValueSubsetMap>;
-  // using SubsetSelectorFallbackParamsRef = std::reference_wrapper<SubsetSelectorFallbackParams>;
-
-  // class LoadBalancerContextWrapper : public LoadBalancerContext {
-  // public:
-  //   LoadBalancerContextWrapper(LoadBalancerContext* wrapped,
-  //                              const std::set<std::string>& filtered_metadata_match_criteria_names);
-  //
-  //   // LoadBalancerContext
-  //   absl::optional<uint64_t> computeHashKey() override { return wrapped_->computeHashKey(); }
-  //   const Router::MetadataMatchCriteria* metadataMatchCriteria() override {
-  //     return metadata_match_.get();
-  //   }
-  //   const Network::Connection* downstreamConnection() const override {
-  //     return wrapped_->downstreamConnection();
-  //   }
-  //   const Http::RequestHeaderMap* downstreamHeaders() const override {
-  //     return wrapped_->downstreamHeaders();
-  //   }
-  //   const HealthyAndDegradedLoad& determinePriorityLoad(
-  //       const PrioritySet& priority_set, const HealthyAndDegradedLoad& original_priority_load,
-  //       const Upstream::RetryPriority::PriorityMappingFunc& priority_mapping_func) override {
-  //     return wrapped_->determinePriorityLoad(priority_set, original_priority_load,
-  //                                            priority_mapping_func);
-  //   }
-  //   bool shouldSelectAnotherHost(const Host& host) override {
-  //     return wrapped_->shouldSelectAnotherHost(host);
-  //   }
-  //   uint32_t hostSelectionRetryCount() const override {
-  //     return wrapped_->hostSelectionRetryCount();
-  //   }
-  //   Network::Socket::OptionsSharedPtr upstreamSocketOptions() const override {
-  //     return wrapped_->upstreamSocketOptions();
-  //   }
-  //   Network::TransportSocketOptionsSharedPtr upstreamTransportSocketOptions() const override {
-  //     return wrapped_->upstreamTransportSocketOptions();
-  //   }
-  //
-  // private:
-  //   LoadBalancerContext* wrapped_;
-  //   Router::MetadataMatchCriteriaConstPtr metadata_match_;
-  // };
-
-  struct SubsetSelectorFallbackParams {
-    envoy::config::cluster::v3::Cluster::LbShuffleSubsetConfig::LbSubsetSelector::
-        LbShuffleSubsetSelectorFallbackPolicy fallback_policy_;
-    const std::set<std::string>* fallback_keys_subset_ = nullptr;
-  };
-
-  struct SubsetSelectorMap {
-    absl::node_hash_map<std::string, SubsetSelectorMapPtr> subset_keys_;
-    SubsetSelectorFallbackParams fallback_params_;
-  };
-
-  // Entry in the subset hierarchy.
-  // class LbShuffleSubsetEntry {
-  // public:
-  //   LbShuffleSubsetEntry() = default;
-  //
-  //   bool initialized() const { return priority_subset_ != nullptr; }
-  //   bool active() const { return initialized() && !priority_subset_->empty(); }
-  //   bool hasChildren() const { return !children_.empty(); }
-  //
-  //   LbSubsetMap children_;
-  //
-  //   // Only initialized if a match exists at this level.
-  //   PriorityShuffleSubsetImplPtr priority_subset_;
-  // };
-
-  // Create filtered default subset (if necessary) and other subsets based on current hosts.
-  void refreshSubsets();
-  void refreshSubsets(uint32_t priority);
-
-  // Called by HostSet::MemberUpdateCb
-  void update(uint32_t priority, const HostVector& hosts_added, const HostVector& hosts_removed);
-
-  // Rebuild the map for single_host_per_subset mode.
-  // void rebuildSingle();
-
-  void updateFallbackSubset(uint32_t priority, const HostVector& hosts_added,
-                            const HostVector& hosts_removed);
-  // void
-  // processSubsets(const HostVector& hosts_added, const HostVector& hosts_removed,
-  //                std::function<void(LbShuffleSubsetEntryPtr)> update_cb,
-  //                std::function<void(LbShuffleSubsetEntryPtr, HostPredicate, const SubsetMetadata&)> cb);
-
-  // HostConstSharedPtr tryChooseHostFromContext(LoadBalancerContext* context, bool& host_chosen);
-  // HostConstSharedPtr
-  // tryChooseHostFromMetadataMatchCriteriaSingle(const Router::MetadataMatchCriteria& match_criteria,
-  //                                              bool& host_chosen);
-
-  // absl::optional<SubsetSelectorFallbackParamsRef>
-  // tryFindSelectorFallbackParams(LoadBalancerContext* context);
-
-  // bool hostMatches(const SubsetMetadata& kvs, const Host& host);
-
-  // LbShuffleSubsetEntryPtr
-  // findSubset(const std::vector<Router::MetadataMatchCriterionConstSharedPtr>& matches);
-
-  // LbShuffleSubsetEntryPtr findOrCreateSubset(LbSubsetMap& subsets, const SubsetMetadata& kvs,
-  //                                     uint32_t idx);
-  // void forEachSubset(LbSubsetMap& subsets, std::function<void(LbShuffleSubsetEntryPtr)> cb);
-  // void purgeEmptySubsets(LbSubsetMap& subsets);
-
-  // std::vector<SubsetMetadata> extractSubsetMetadata(const std::set<std::string>& subset_keys,
-  //                                                   const Host& host);
-  // std::string describeMetadata(const SubsetMetadata& kvs);
-
-  const LoadBalancerType lb_type_;
-  const absl::optional<envoy::config::cluster::v3::Cluster::RingHashLbConfig> lb_ring_hash_config_;
-  const absl::optional<envoy::config::cluster::v3::Cluster::MaglevLbConfig> lb_maglev_config_;
-  const absl::optional<envoy::config::cluster::v3::Cluster::LeastRequestLbConfig>
-      least_request_config_;
-  const envoy::config::cluster::v3::Cluster::CommonLbConfig common_config_;
-  ClusterStats& stats_;
-  Stats::Scope& scope_;
-  Runtime::Loader& runtime_;
-  Random::RandomGenerator& random_;
-
-  // const envoy::config::cluster::v3::Cluster::LbShuffleSubsetConfig::LbSubsetFallbackPolicy
-  //     fallback_policy_;
-  const uint32_t shard_size_;
-
-  // const SubsetMetadata default_subset_metadata_;
-  // std::vector<ShuffleSubsetSelectorPtr> subset_selectors_;
-
-  const PrioritySet& original_priority_set_;
-  const PrioritySet* original_local_priority_set_;
-  // Common::CallbackHandle* original_priority_set_callback_handle_;
-
-  // LbShuffleSubsetEntryPtr subset_any_;
-  // LbShuffleSubsetEntryPtr fallback_subset_;
-  PriorityShuffleSubsetImplPtr priority_subset_;
-  // LbShuffleSubsetEntryPtr panic_mode_subset_;
-
-  // LbShuffleSubsetEntryPtr selector_fallback_subset_default_;
-
-  // Forms a trie-like structure. Requires lexically sorted Host and Route metadata.
-  // LbSubsetMap subsets_;
-  // Forms a trie-like structure of lexically sorted keys+fallback policy from subset
-  // selectors configuration
-  // SubsetSelectorMapPtr selectors_;
-
-  // std::string single_key_;
-  // absl::flat_hash_map<HashedValue, HostConstSharedPtr> single_host_per_subset_map_;
-  // Stats::Gauge* single_duplicate_stat_{};
-
-  // const bool locality_weight_aware_;
-  // const bool scale_locality_weight_;
-  // const bool list_as_any_;
-
-  friend class ShuffleSubsetLoadBalancerDescribeMetadataTester;
-};
-
 } // namespace Upstream
 } // namespace Envoy
